refactor scripture to create a data struct
this refactor seperates the scripture data struct from the list view items.
this makes it easier to update, manipulate sync with things like database and
external sources for importing scripture
